+++ frameworks/qwik/tsdown.config.ts
@@ -2,12 +2,26 @@ import { defineConfig } from 'tsdown';
 
 export default defineConfig({
   entry: ['./src/index.ts'],
+  external: ['@formisch/core/qwik'],
   outDir: 'dist',
   outExtensions: () => ({
-    dts: '.d.ts',
+    dts: '.qwik.d.ts',
     js: '.qwik.js',
   }),
-  dts: {
-    resolve: ['@formisch/core/qwik', '@formisch/methods/qwik'],
+  outputOptions: {
+    paths: { '@formisch/core/qwik': './internals.qwik.js' },
   },
+  dts: {
+    resolve: ['@formisch/methods/qwik'],
+  },
+  copy: [
+    {
+      from: '../../packages/core/dist/index.qwik.js',
+      to: 'dist/internals.qwik.js',
+    },
+    {
+      from: '../../packages/core/dist/index.qwik.d.ts',
+      to: 'dist/internals.qwik.d.ts',
+    },
+  ],
 });